- Move all MCG helpers to the general helpers file - Add bucket with objects deletion test - Add assertion messages
