+++ tests/manage/mcg/test_multi_region.py
@@ -10,7 +10,7 @@ from ocs_ci.ocs import constants
 from ocs_ci.ocs.constants import BS_AUTH_FAILED, BS_OPTIMAL
 from ocs_ci.ocs.exceptions import TimeoutExpiredError
 from ocs_ci.utility.utils import TimeoutSampler
-from tests.manage.mcg.helpers import retrieve_test_objects_to_pod, sync_object_directory
+from tests.helpers import retrieve_test_objects_to_pod, sync_object_directory
 from tests.helpers import verify_s3_object_integrity
 
 logger = logging.getLogger(__name__)
